pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pedigree-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",   # needed by FastAPI TestClient
    "pytest>=8.0",
    "pytest-playwright>=0.5",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "ui: Playwright browser tests (running app + playwright install chromium)",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "httpx>=0.27",
]
